azure-bot/requirements.txt
azure-servicebus
botbuilder-core
aiohttp>=3.0.0
uamqp
